Refactor funding
