+++ tools/fund.js
@@ -1,13 +1,19 @@
 import { program } from "commander";
 import { ethers } from "ethers";
 
+import { buildMimcHasher } from "../circuits/util.js";
 import deployments from "../contracts/deployments.json" with { type: "json" };
 
 const options = program
   .name("shadowlings-fund")
-  .option("--token <address>", "The token to mint")
-  .requiredOption("--to <address>", "The amount to transfer")
+  .requiredOption("--owner <address>", "The owner of the shadowling")
+  .option(
+    "--entropy <value>",
+    "Additional entropy for preserving privacy with recovery",
+  )
+  .requiredOption("--salt <value>", "The shadowling specific salt")
   .requiredOption("--amount <value>", "The recipient of the transfer")
+  .option("--token <address>", "The token to mint")
   .option("--demo", "Run in demo mode, funding with hardcoded contracts.")
   .option("--rpc-url <url>", "The node RPC to connect to")
   .parse()
@@ -19,42 +25,12 @@ async function main() {
   );
   const signer = await provider.getSigner();
 
-  const to = ethers.getAddress(options.to);
-  const value = ethers.parseEther(options.amount);
-
   const { chainId } = await provider.getNetwork();
-  let tokens = [ethers.getAddress(options.token ?? ethers.ZeroAddress)];
-  if (options.demo) {
-    tokens = [
-      ...new Set([
-        ...tokens,
-        ethers.ZeroAddress,
-        deployments[chainId].ShadowToken,
-      ]),
-    ];
-  }
-
-  for (const token of tokens) {
-    let transactionData;
-    if (token === ethers.ZeroAddress) {
-      transactionData = { to, value };
-    } else {
-      transactionData = {
-        to: token,
-        data: TOKEN.encodeFunctionData("mint", [to, value]),
-      };
-    }
-
-    const transaction = await signer.sendTransaction(transactionData);
-    const receipt = await transaction.wait();
-
-    console.log(receipt);
-  }
-
   const shadowlings = new ethers.Contract(
     options.shadowlings ?? deployments[chainId].Shadowlings,
     [
       `function ENTRY_POINT() view returns (address)`,
+      `function getShadowling(uint256 commit) view returns (address)`,
     ],
     provider,
   );
@@ -67,8 +43,54 @@ async function main() {
     signer,
   );
 
-  if (await entryPoint.balanceOf(to) < ethers.parseEther("1.0")) {
-    const transaction = await entryPoint.depositTo(to, {
+  const owner = ethers.getAddress(options.owner);
+  const entropy = options.entropy
+    ? BigInt(ethers.hexlify(ethers.toUtf8Bytes(options.entropy)))
+    : 0n;
+  const salt = BigInt(options.salt);
+  const pepper = 42;
+
+  const mimc = await buildMimcHasher();
+
+  const ownerHash = mimc(owner, entropy);
+  const saltHash = mimc(salt, pepper);
+  const commit = mimc(ownerHash, saltHash);
+  console.log({ entropy, commit });
+
+  const shadowling = await shadowlings.getShadowling(commit);
+  console.log({ shadowling });
+
+  let tokens = [ethers.getAddress(options.token ?? ethers.ZeroAddress)];
+  if (options.demo) {
+    tokens = [
+      ...new Set([
+        ...tokens,
+        ethers.ZeroAddress,
+        deployments[chainId].ShadowToken,
+      ]),
+    ];
+  }
+  const value = ethers.parseEther(options.amount);
+
+  for (const token of tokens) {
+    let transactionData;
+    if (token === ethers.ZeroAddress) {
+      transactionData = { to: shadowling, value };
+    } else {
+      transactionData = {
+        to: token,
+        data: TOKEN.encodeFunctionData("mint", [shadowling, value]),
+      };
+    }
+
+    const transaction = await signer.sendTransaction(transactionData);
+    const receipt = await transaction.wait();
+
+    console.log(receipt);
+  }
+
+  if (await entryPoint.balanceOf(shadowling) < ethers.parseEther("1.0")) {
+    const transaction = await entryPoint.depositTo(shadowling, {
       value: ethers.parseEther("10.0"),
     });
     const receipt = await transaction.wait();
